requirements.txt
transformers
datasets
peft
accelerate
pillow
requests
bitsandbytes
